--- packages/pglite/serve.json
@@ -1,17 +1,0 @@
-{
-  "headers": [
-    {
-      "source": "**/*",
-      "headers": [
-        {
-          "key": "Cross-Origin-Embedder-Policy",
-          "value": "require-corp"
-        },
-        {
-          "key": "Cross-Origin-Opener-Policy",
-          "value": "same-origin"
-        }
-      ]
-    }
-  ]
-}
